Start splitting out FITS-WCS-specific code into separate file
