+++ astropy/visualization/wcsaxes/utils.py
@@ -119,27 +119,6 @@ def get_coord_meta(frame):
     return coord_meta
 
 
-def coord_type_from_ctype(ctype):
-    """
-    Determine whether a particular WCS ctype corresponds to an angle or scalar
-    coordinate.
-    """
-    if ctype[:4] == 'RA--':
-        return 'longitude', u.hourangle, None
-    elif ctype[:4] == 'HPLN':
-        return 'longitude', u.arcsec, 180.
-    elif ctype[:4] == 'HPLT':
-        return 'latitude', u.arcsec, None
-    elif ctype[:4] == 'HGLN':
-        return 'longitude', None, 180.
-    elif ctype[1:4] == 'LON' or ctype[2:4] == 'LN':
-        return 'longitude', None, None
-    elif ctype[:4] == 'DEC-' or ctype[1:4] == 'LAT' or ctype[2:4] == 'LT':
-        return 'latitude', None, None
-    else:
-        return 'scalar', None, None
-
-
 def transform_contour_set_inplace(cset, transform):
     """
     Transform a contour set in-place using a specified
